test: table 관련 테스트 분리
